Update test fixtures: ensure rollback of any transactions in the test DB session
